fix: Updated GitHub spec. Fixed issues with it.

+++ src/tests/AutoSDK.SnapshotTests/Snapshots/Filtering/SystemTextJson/_#G.Models.Migration.g.verified.cs
@@ -10,11 +10,11 @@ namespace G
     public sealed partial class Migration
     {
         /// <summary>
-        /// Example: 79
+        /// Example: 79L
         /// </summary>
         [global::System.Text.Json.Serialization.JsonPropertyName("id")]
         [global::System.Text.Json.Serialization.JsonRequired]
-        public required int Id { get; set; }
+        public required long Id { get; set; }
 
         /// <summary>
         /// A GitHub user.
