+++ src/tests/AutoSDK.SnapshotTests/Snapshots/GitHub/NewtonsoftJson/_#G.IReactionsClient.ReactionsDeleteForIssueComment.g.verified.cs
@@ -7,7 +7,8 @@ namespace G
     {
         /// <summary>
         /// Delete an issue comment reaction<br/>
-        /// **Note:** You can also specify a repository by `repository_id` using the route `DELETE delete /repositories/:repository_id/issues/comments/:comment_id/reactions/:reaction_id`.<br/>
+        /// &gt; [!NOTE]<br/>
+        /// &gt; You can also specify a repository by `repository_id` using the route `DELETE delete /repositories/:repository_id/issues/comments/:comment_id/reactions/:reaction_id`.<br/>
         /// Delete a reaction to an [issue comment](https://docs.github.com/rest/issues/comments#get-an-issue-comment).
         /// </summary>
         /// <param name="owner"></param>
@@ -19,7 +20,7 @@ namespace G
         global::System.Threading.Tasks.Task ReactionsDeleteForIssueCommentAsync(
             string owner,
             string repo,
-            int commentId,
+            long commentId,
             int reactionId,
             global::System.Threading.CancellationToken cancellationToken = default);
     }
