+++ src/tests/AutoSDK.SnapshotTests/Snapshots/GitHub/NewtonsoftJson/_#G.Models.Deployment.g.verified.cs
@@ -19,10 +19,10 @@ namespace G
 
         /// <summary>
         /// Unique identifier of the deployment<br/>
-        /// Example: 42
+        /// Example: 42L
         /// </summary>
         [global::Newtonsoft.Json.JsonProperty("id", Required = global::Newtonsoft.Json.Required.Always)]
-        public int Id { get; set; } = default!;
+        public long Id { get; set; } = default!;
 
         /// <summary>
         /// Example: MDEwOkRlcGxveW1lbnQx
