+++ src/tests/AutoSDK.SnapshotTests/Snapshots/GitHub/NewtonsoftJson/_#G.Models.Migration.g.verified.cs
@@ -10,10 +10,10 @@ namespace G
     public sealed partial class Migration
     {
         /// <summary>
-        /// Example: 79
+        /// Example: 79L
         /// </summary>
         [global::Newtonsoft.Json.JsonProperty("id", Required = global::Newtonsoft.Json.Required.Always)]
-        public int Id { get; set; } = default!;
+        public long Id { get; set; } = default!;
 
         /// <summary>
         /// A GitHub user.
