+++ src/tests/AutoSDK.SnapshotTests/Snapshots/GitHub/NewtonsoftJson/_#G.Models.PrivateUser.g.verified.cs
@@ -16,10 +16,10 @@ namespace G
         public string Login { get; set; } = default!;
 
         /// <summary>
-        /// Example: 1
+        /// Example: 1L
         /// </summary>
         [global::Newtonsoft.Json.JsonProperty("id", Required = global::Newtonsoft.Json.Required.Always)]
-        public int Id { get; set; } = default!;
+        public long Id { get; set; } = default!;
 
         /// <summary>
         /// Example: MDQ6VXNlcjE=
