+++ src/tests/AutoSDK.SnapshotTests/Snapshots/GitHub/NewtonsoftJson/_#G.Models.ReviewComment.g.verified.cs
@@ -16,16 +16,16 @@ namespace G
         public string Url { get; set; } = default!;
 
         /// <summary>
-        /// Example: 42
+        /// Example: 42L
         /// </summary>
         [global::Newtonsoft.Json.JsonProperty("pull_request_review_id", Required = global::Newtonsoft.Json.Required.Always)]
-        public int? PullRequestReviewId { get; set; } = default!;
+        public long? PullRequestReviewId { get; set; } = default!;
 
         /// <summary>
-        /// Example: 10
+        /// Example: 10L
         /// </summary>
         [global::Newtonsoft.Json.JsonProperty("id", Required = global::Newtonsoft.Json.Required.Always)]
-        public int Id { get; set; } = default!;
+        public long Id { get; set; } = default!;
 
         /// <summary>
         /// Example: MDI0OlB1bGxSZXF1ZXN0UmV2aWV3Q29tbWVudDEw
