+++ src/tests/AutoSDK.SnapshotTests/Snapshots/GitHub/NewtonsoftJson/_#G.ReactionsClient.ReactionsCreateForPullRequestReviewComment.g.verified.cs
@@ -10,14 +10,14 @@ namespace G
             global::System.Net.Http.HttpClient httpClient,
             ref string owner,
             ref string repo,
-            ref int commentId,
+            ref long commentId,
             global::G.ReactionsCreateForPullRequestReviewCommentRequest request);
         partial void PrepareReactionsCreateForPullRequestReviewCommentRequest(
             global::System.Net.Http.HttpClient httpClient,
             global::System.Net.Http.HttpRequestMessage httpRequestMessage,
             string owner,
             string repo,
-            int commentId,
+            long commentId,
             global::G.ReactionsCreateForPullRequestReviewCommentRequest request);
         partial void ProcessReactionsCreateForPullRequestReviewCommentResponse(
             global::System.Net.Http.HttpClient httpClient,
@@ -41,7 +41,7 @@ namespace G
         public async global::System.Threading.Tasks.Task<global::G.Reaction> ReactionsCreateForPullRequestReviewCommentAsync(
             string owner,
             string repo,
-            int commentId,
+            long commentId,
             global::G.ReactionsCreateForPullRequestReviewCommentRequest request,
             global::System.Threading.CancellationToken cancellationToken = default)
         {
@@ -133,7 +133,7 @@ namespace G
         public async global::System.Threading.Tasks.Task<global::G.Reaction> ReactionsCreateForPullRequestReviewCommentAsync(
             string owner,
             string repo,
-            int commentId,
+            long commentId,
             global::G.ReactionsCreateForPullRequestReviewCommentRequestContent content,
             global::System.Threading.CancellationToken cancellationToken = default)
         {
