+++ src/tests/AutoSDK.SnapshotTests/Snapshots/GitHub/SystemTextJson/_#G.GistsClient.GistsDeleteComment.g.verified.cs
@@ -9,12 +9,12 @@ namespace G
         partial void PrepareGistsDeleteCommentArguments(
             global::System.Net.Http.HttpClient httpClient,
             ref string gistId,
-            ref int commentId);
+            ref long commentId);
         partial void PrepareGistsDeleteCommentRequest(
             global::System.Net.Http.HttpClient httpClient,
             global::System.Net.Http.HttpRequestMessage httpRequestMessage,
             string gistId,
-            int commentId);
+            long commentId);
         partial void ProcessGistsDeleteCommentResponse(
             global::System.Net.Http.HttpClient httpClient,
             global::System.Net.Http.HttpResponseMessage httpResponseMessage);
@@ -33,7 +33,7 @@ namespace G
         /// <exception cref="global::System.InvalidOperationException"></exception>
         public async global::System.Threading.Tasks.Task<global::G.BasicError> GistsDeleteCommentAsync(
             string gistId,
-            int commentId,
+            long commentId,
             global::System.Threading.CancellationToken cancellationToken = default)
         {
             PrepareArguments(
